--- a/packages/dds/tree/src/core/index.ts
+++ b/packages/dds/tree/src/core/index.ts
@@ -59,6 +59,7 @@ export {
     keyAsDetachedField,
     visitDelta,
     setGenericTreeField,
+    rootFieldKeySymbol,
 } from "../tree";
 
 export {