--- a/packages/dds/tree/src/feature-libraries/editable-tree/editableTree.ts
+++ b/packages/dds/tree/src/feature-libraries/editable-tree/editableTree.ts
@@ -11,7 +11,6 @@ import {
     symbolIsFieldKey,
     IEditableForest,
     TreeNavigationResult,
-    mapCursorField,
     ITreeSubscriptionCursor,
     ITreeSubscriptionCursorState,
     FieldSchema,
@@ -20,6 +19,8 @@ import {
     NamedTreeSchema,
     ValueSchema,
     lookupTreeSchema,
+    mapCursorFieldNew,
+    mapCursorFields,
 } from "../../core";
 import { brand } from "../../util";
 import { FieldKind, Multiplicity } from "../modular-schema";
@@ -193,15 +194,21 @@ export class ProxyTarget {
         key?: FieldKey,
         nameOnly = true,
     ): NamedTreeSchema | TreeSchemaIdentifier | undefined {
-        let typeName = this.cursor.type;
+        let typeName: TreeSchemaIdentifier | undefined = this.cursor.type;
         if (key !== undefined) {
+            // TODO: remove option to use this for getting field types:
+            // Once fields are properly wrapped, get the field, and get its type.
+
             const fieldKind = this.lookupFieldKind(key);
             if (fieldKind.multiplicity === Multiplicity.Sequence) {
                 return undefined;
             }
-            const fieldLength = this.cursor.childFieldLength(key);
-            assert(fieldLength <= 1, 0x3c5 /* invalid non sequence */);
-            typeName = mapCursorField(this.cursor, key, (c) => c.type)[0];
+
+            this.cursor.enterField(key);
+            const types = mapCursorFieldNew(this.cursor, (c) => c.type);
+            this.cursor.exitField();
+            assert(types.length <= 1, 0x3c5 /* invalid non sequence */);
+            typeName = types[0];
         }
         if (nameOnly) {
             return typeName;
@@ -230,20 +237,15 @@ export class ProxyTarget {
     }
 
     public getFieldKeys(): FieldKey[] {
-        // For now this is an approximation:
-        const fieldKeys: FieldKey[] = [];
-        for (const key of this.cursor.keys) {
-            // TODO: with new cursor API, field iteration will skip empty fields and this check can be removed.
-            if (this.has(key)) {
-                fieldKeys.push(key);
-            }
-        }
-        return fieldKeys;
+        return mapCursorFields(this.cursor, (c) => c.getFieldKey());
     }
 
     public has(field: FieldKey): boolean {
         // Make fields present only if non-empty.
-        return this.cursor.childFieldLength(field) !== 0;
+        this.cursor.enterField(field);
+        const length = this.cursor.getFieldLength();
+        this.cursor.exitField();
+        return length !== 0;
     }
 
     /**
@@ -271,11 +273,12 @@ export class ProxyTarget {
             this.getType(undefined, false) as NamedTreeSchema,
         );
         // Make the childTargets:
-        const childTargets = mapCursorField(
+        this.cursor.enterField(field);
+        const childTargets = mapCursorFieldNew(
             this.cursor,
-            field,
             (c) => new ProxyTarget(this.context, c),
         );
+        this.cursor.exitField();
         return proxifyField(fieldSchema, field, childTargets, unwrap);
     }
 
@@ -422,11 +425,12 @@ function inProxyOrUnwrap(target: ProxyTarget, unwrap: boolean): UnwrappedEditabl
         }
         const primary = target.getPrimaryArrayKey();
         if (primary !== undefined) {
-            const childTargets = mapCursorField(
+            target.cursor.enterField(primary.key);
+            const childTargets = mapCursorFieldNew(
                 target.cursor,
-                primary.key,
                 (c) => new ProxyTarget(target.context, c),
             );
+            target.cursor.exitField();
             return childTargets.map((childTarget) => inProxyOrUnwrap(childTarget, unwrap));
         }
     }