--- a/packages/dds/tree/src/feature-libraries/forestIndex.ts
+++ b/packages/dds/tree/src/feature-libraries/forestIndex.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { assert, bufferToString, IsoBuffer } from "@fluidframework/common-utils";
+import { bufferToString, IsoBuffer } from "@fluidframework/common-utils";
 import { IFluidHandle } from "@fluidframework/core-interfaces";
 import {
     IFluidDataStoreRuntime,
@@ -29,9 +29,10 @@ import {
     recordDependency,
     JsonableTree,
     Delta,
+    mapCursorFieldNew,
 } from "../core";
-import { jsonableTreeFromCursor } from "./treeTextCursorLegacy";
 import { singleTextCursor } from "./treeTextCursor";
+import { jsonableTreeFromCursorNew } from ".";
 
 /**
  * The storage key for the blob in the summary containing tree data
@@ -88,14 +89,12 @@ export class ForestIndex implements Index<unknown>, SummaryElement {
         // TODO: maybe assert there are no other roots
         // (since we don't save them, and they should not exist outside transactions).
         const rootAnchor = this.forest.root(this.forest.rootField);
-        const roots: JsonableTree[] = [];
-        let result = this.forest.tryMoveCursorTo(rootAnchor, this.cursor);
-        while (result === TreeNavigationResult.Ok) {
-            roots.push(jsonableTreeFromCursor(this.cursor));
-            result = this.cursor.seek(1);
-        }
+        const result = this.forest.tryMoveCursorTo(rootAnchor, this.cursor);
+        const roots =
+            result === TreeNavigationResult.Ok
+                ? mapCursorFieldNew(this.cursor, jsonableTreeFromCursorNew)
+                : [];
         this.cursor.clear();
-        assert(result === TreeNavigationResult.NotFound, 0x34b /* Unsupported navigation result */);
 
         return JSON.stringify(roots);
     }