--- a/packages/dds/tree/src/test/cursor.spec.ts
+++ b/packages/dds/tree/src/test/cursor.spec.ts
@@ -12,6 +12,7 @@ import {
     ITreeCursorNew as ITreeCursor,
     mapCursorFields,
     CursorLocationType,
+    rootFieldKey,
 } from "../tree";
 import { brand } from "../util";
 
@@ -100,6 +101,7 @@ export function testJsonableTreeCursor(
     cursorName: string,
     factory: (data: JsonableTree) => ITreeCursor,
     dataFromCursor: (cursor: ITreeCursor) => JsonableTree,
+    testRooted = true,
 ): void {
     describe(`${cursorName} cursor implementation`, () => {
         describe("extract roundtrip", () => {
@@ -116,10 +118,13 @@ export function testJsonableTreeCursor(
             }
         });
 
-        it("up from root", () => {
-            const cursor = factory({ type: brand("Foo") });
-            assert.throws(() => cursor.exitNode());
-        });
+        // TODO: revisit spec for forest cursors and root and clarify what should be tested for them regarding Up from root.
+        if (testRooted) {
+            it("up from root", () => {
+                const cursor = factory({ type: brand("Foo") });
+                assert.throws(() => cursor.exitNode());
+            });
+        }
 
         describe("keys", () => {
             const getFieldKey = (cursor: ITreeCursor) => cursor.getFieldKey();
@@ -423,6 +428,13 @@ export function testJsonableTreeCursor(
         });
 
         describe("getPath() returns correct path for", () => {
+            const parent = testRooted
+                ? undefined
+                : {
+                      parent: undefined,
+                      parentField: rootFieldKey,
+                      parentIndex: 0,
+                  };
             it(`first node in a root trait`, () => {
                 const cursor = factory({
                     type: brand("Foo"),
@@ -431,7 +443,7 @@ export function testJsonableTreeCursor(
                 cursor.enterField(brand("key"));
                 cursor.firstNode();
                 assert.deepEqual(cursor.getPath(), {
-                    parent: undefined,
+                    parent,
                     parentField: brand<FieldKey>("key"),
                     parentIndex: 0,
                 });
@@ -450,7 +462,7 @@ export function testJsonableTreeCursor(
                 cursor.enterField(brand("key"));
                 cursor.enterNode(1);
                 assert.deepEqual(cursor.getPath(), {
-                    parent: undefined,
+                    parent,
                     parentField: brand<FieldKey>("key"),
                     parentIndex: 1,
                 });
@@ -474,7 +486,7 @@ export function testJsonableTreeCursor(
                 assert.equal(cursor.firstNode(), true);
                 assert.deepEqual(cursor.getPath(), {
                     parent: {
-                        parent: undefined,
+                        parent,
                         parentField: brand<FieldKey>("a"),
                         parentIndex: 0,
                     },