--- a/packages/dds/tree/src/test/feature-libraries/treeTextCursor.spec.ts
+++ b/packages/dds/tree/src/test/feature-libraries/treeTextCursor.spec.ts
@@ -2,7 +2,16 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
-import { jsonableTreeFromCursorNew, singleTextCursorNew } from "../../feature-libraries";
+import { strict as assert } from "assert";
+import { initializeForest, InMemoryStoredSchemaRepository, TreeNavigationResult } from "../../core";
+import { jsonSchemaData } from "../../domains";
+import {
+    defaultSchemaPolicy,
+    jsonableTreeFromCursorNew,
+    ObjectForest,
+    singleTextCursorNew,
+} from "../../feature-libraries";
+import { ITreeCursorNew } from "../../tree";
 import { cursorTestCases, testCursors, testJsonableTreeCursor } from "../cursor.spec";
 
 // Tests for TextCursor and jsonableTreeFromCursor.
@@ -10,23 +19,24 @@ import { cursorTestCases, testCursors, testJsonableTreeCursor } from "../cursor.
 // since its used in the below test suite to test other formats.
 testJsonableTreeCursor("textTreeFormat", singleTextCursorNew, jsonableTreeFromCursorNew);
 
-// TODO make object forest cursor compatible with new API
 // TODO: put these in a better place / unify with object forest tests.
-// testJsonCompatibleCursor(
-//     "object-forest cursor",
-//     (data): ITreeCursor => {
-//         const schemaData: SchemaData = {
-//             globalFieldSchema: new Map(),
-//             treeSchema: jsonTypeSchema,
-//         };
-//         const forest = new ObjectForest(new StoredSchemaRepository(defaultSchemaPolicy, schemaData));
-//         initializeForest(forest, [singleTextCursorNew(data)]);
-//         const cursor = forest.allocateCursor();
-//         assert.equal(forest.tryMoveCursorTo(forest.root(forest.rootField), cursor), TreeNavigationResult.Ok);
-//         return cursor;
-//     },
-//     jsonableTreeFromCursor,
-// );
+testJsonableTreeCursor(
+    "object-forest cursor",
+    (data): ITreeCursorNew => {
+        const forest = new ObjectForest(
+            new InMemoryStoredSchemaRepository(defaultSchemaPolicy, jsonSchemaData),
+        );
+        initializeForest(forest, [singleTextCursorNew(data)]);
+        const cursor = forest.allocateCursor();
+        assert.equal(
+            forest.tryMoveCursorTo(forest.root(forest.rootField), cursor),
+            TreeNavigationResult.Ok,
+        );
+        return cursor;
+    },
+    jsonableTreeFromCursorNew,
+    false,
+);
 
 testCursors(
     "textTreeFormat",