--- a/packages/dds/tree/src/test/feature-libraries/treeTextCursorLegacy.spec.ts
+++ b/packages/dds/tree/src/test/feature-libraries/treeTextCursorLegacy.spec.ts
@@ -3,18 +3,12 @@
  * Licensed under the MIT License.
  */
 
-import { strict as assert } from "assert";
-import { jsonSchemaData } from "../../domains";
-import { defaultSchemaPolicy, ObjectForest, singleTextCursorNew } from "../../feature-libraries";
-
 // Allow importing from this specific file which is being tested:
 import {
     jsonableTreeFromCursor,
     singleTextCursor,
     // eslint-disable-next-line import/no-internal-modules
 } from "../../feature-libraries/treeTextCursorLegacy";
-import { initializeForest, ITreeCursor, TreeNavigationResult } from "../../forest";
-import { InMemoryStoredSchemaRepository } from "../../schema-stored";
 import { cursorTestCases } from "../cursor.spec";
 import { testCursors, testJsonableTreeCursor } from "../cursorLegacy.spec";
 
@@ -22,24 +16,6 @@ import { testCursors, testJsonableTreeCursor } from "../cursorLegacy.spec";
 // Checks to make sure singleTextCursor and test datasets are working properly,
 // since its used in the below test suite to test other formats.
 testJsonableTreeCursor("textTreeFormat", singleTextCursor, jsonableTreeFromCursor);
-
-// TODO: put these in a better place / unify with object forest tests.
-testJsonableTreeCursor(
-    "object-forest cursor",
-    (data): ITreeCursor => {
-        const forest = new ObjectForest(
-            new InMemoryStoredSchemaRepository(defaultSchemaPolicy, jsonSchemaData),
-        );
-        initializeForest(forest, [singleTextCursorNew(data)]);
-        const cursor = forest.allocateCursor();
-        assert.equal(
-            forest.tryMoveCursorTo(forest.root(forest.rootField), cursor),
-            TreeNavigationResult.Ok,
-        );
-        return cursor;
-    },
-    jsonableTreeFromCursor,
-);
 
 testCursors(
     "textTreeFormat",