--- a/packages/dds/tree/src/tree/index.ts
+++ b/packages/dds/tree/src/tree/index.ts
@@ -48,6 +48,7 @@ export {
     keyAsDetachedField,
     rootFieldKey,
     NodeData,
+    rootFieldKeySymbol,
 } from "./types";
 export { DeltaVisitor, visitDelta } from "./visitDelta";