## Description

Migrates forest to new cursor API.

This mainly impacts object-forest (which has to implement the new API), and editable-tree (which has to consume the new API).

Other than migrating to the new API, there are some significant design changes in object-forest:
1. It now uses MapTree instead of JsonableTree
2. It has a dummy root node above the detached sequences which simplifies apply delta and some other aspects of root handling.
3. Its cursor wraps an existing cursor instead of extending one.

## Performance:
Faster, though one regression:

Before
```
status  name                                                     period (ns/op)  relative margin of error  iterations  samples  total time (s)
------  -------------------------------------------------------  --------------  ------------------------  ----------  -------  --------------
    ✔   cursorToJsonObject(object-forest Cursor): 'canada'         45,195,222.5                    ±1.81%          82       41            5.79
    ✔   jsonableTreeFromCursor(object-forest Cursor): 'canada'     96,763,088.4                    ±3.05%          30       30            5.94
    ✔   sum(object-forest Cursor): 'canada'                        48,716,704.3                    ±0.22%          76       38            5.84
    ✔   averageTwoValues(object-forest Cursor): 'canada'           36,658,514.6                    ±1.54%          98       49            5.64
    ✔   cursorToJsonObject(object-forest Cursor): 'twitter'        41,951,789.8                    ±1.16%          86       43            5.74
    ✔   jsonableTreeFromCursor(object-forest Cursor): 'twitter'    80,564,255.5                    ±3.16%          35       35            5.77
    ✔   sum(object-forest Cursor): 'twitter'                       38,948,035.3                    ±0.13%          92       46            5.67
    ✔   averageTwoValues(object-forest Cursor): 'twitter'             302,512.5                    ±0.94%      16,245       95            5.43
```

After:
```
status  name                                                     period (ns/op)  relative margin of error  iterations  samples  total time (s)
------  -------------------------------------------------------  --------------  ------------------------  ----------  -------  --------------
    ✔   cursorToJsonObject(object-forest Cursor): 'canada'         53,977,035.4                    ±1.02%          49       49            5.59
    ✔   jsonableTreeFromCursor(object-forest Cursor): 'canada'     83,449,600.6                    ±3.66%          34       34            5.86
    ✔   sum(object-forest Cursor): 'canada'                        27,261,029.8                    ±0.27%         126       63            5.49
    ✔   averageTwoValues(object-forest Cursor): 'canada'           21,406,211.2                    ±0.27%         183       61            5.63
    ✔   cursorToJsonObject(object-forest Cursor): 'twitter'        32,734,807.8                    ±1.26%         108       54            5.61
    ✔   jsonableTreeFromCursor(object-forest Cursor): 'twitter'    57,881,098.4                    ±2.75%          46       46            5.63
    ✔   sum(object-forest Cursor): 'twitter'                       16,481,119.6                    ±0.52%         252       63            5.50
    ✔   averageTwoValues(object-forest Cursor): 'twitter'             241,298.7                    ±0.45%      20,657       91            5.41

```